clean up names, interfaces and other things
