+++ src/hooks/useNewDeliveryContext.ts
@@ -1,9 +1,6 @@
 import * as React from 'react'
 import { NewDeliveryContext } from '@/contexts/NewDelivery'
-import { NewDeliveryContextValues } from '@/interfaces/newDeliveryContextValues'
 
-export const useNewDeliveryContext = (): NewDeliveryContextValues => {
-  const context = React.useContext(NewDeliveryContext)
-
-  return context
+export const useNewDeliveryContext = () => {
+  return React.useContext(NewDeliveryContext)
 }
