+++ src/hooks/useThemeContext.ts
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { ColorModeContext } from '@/contexts/Theme'
-import { ColorModeContextValue } from '@/interfaces/colorModeContextValue'
+import { ColorModeContextValue } from '@/interfaces/ColorModeContextValue.ts'
 
 export const useThemeContext = (): ColorModeContextValue => {
   const context = React.useContext(ColorModeContext)
