--- src/interfaces/dataTable.ts
@@ -1,6 +1,0 @@
-export interface Column {
-  key: string
-  title: string
-  minWidth?: number
-  align?: 'right' | 'left' | 'center' | 'inherit' | 'justify'
-}
