--- src/interfaces/good.ts
@@ -1,4 +1,0 @@
-export interface Good {
-  goodTypeStep3: string
-  goodQuantityStep3: string
-}
