--- src/interfaces/goodQuantityStep3.ts
@@ -1,5 +1,0 @@
-export interface GoodQuantityStep3 {
-  pallets: number
-  packages: number
-  pieces: number
-}
