--- src/interfaces/moveGood.ts
@@ -1,5 +1,0 @@
-export interface MoveGood {
-  goodTypeStep4: string // palettes | packages | pieces
-  goodQuantityStep4: number
-  zone: string
-}
