+++ src/router/routerPaths.ts
@@ -7,19 +7,11 @@ export const DEFAULTLAYOUT_PATH = '/defaultLayout'
 export const ADMIN_PATH = 'admin'
 export const MAIN_PATH = 'main'
 export const USERS_PATH = '/users'
-// export const ZONES_PATH = 'zones'
 export const ZONES_PATH = '/admin/zones'
-// export const ZONES_CONTENT_PATH = 'zonesContent'
 export const ZONES_CONTENT_PATH = '/main/zonesContent'
-// export const VENDORS_PATH = 'vendors'
 export const VENDORS_PATH = '/admin/vendors'
-// export const ROLES_PATH = 'roles'
 export const ROLES_PATH = '/admin/roles'
-// export const TYPESGOODS_PATH = 'typesGoods'
 export const TYPESGOODS_PATH = '/admin/typesGoods'
-// export const TYPESDIFFERENCE_PATH = 'typesDifference'
 export const TYPESDIFFERENCE_PATH = '/admin/typesDifference'
-// export const MARKERS_PATH = 'markers'
 export const MARKERS_PATH = '/admin/markers'
-// export const DELIVERIES_PATH = 'deliveries'
 export const DELIVERIES_PATH = '/main/deliveries'
